Improve decomposition of OpenAPI and address comments
