+++ resources/openapi-mt.json
@@ -1304,6 +1304,7 @@
         "parameters": [
           {
             "description": "Query to execute (e.g. SQL)",
+            "example": "select 1",
             "in": "query",
             "name": "query",
             "required": true,
