--- src/adapter/http/resources/openapi-validator/config.yaml
@@ -1,5 +1,0 @@
-errorsOnly: true
-# limits:
-#   warnings: 25
-# outputFormat: 'json'
-# summaryOnly: true
