+++ src/adapter/http/resources/openapi/linter-config.yaml
@@ -1,0 +1,1 @@
+errorsOnly: true
