+++ src/adapter/http/resources/openapi/linter-ruleset.yaml
