+++ src/e2e/app/cli/common/src/kamu_api_server_client_ext.rs
@@ -1038,7 +1038,7 @@ impl SwaggerApi<'_> {
     pub async fn main_page(&self) -> String {
         let response = self
             .client
-            .rest_api_call(Method::GET, "/swagger/", None)
+            .rest_api_call(Method::GET, "/openapi", None)
             .await;
 
         pretty_assertions::assert_eq!(StatusCode::OK, response.status());
